+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/navigation/audioguidance/MapboxAudioGuidanceServices.kt
@@ -18,5 +18,5 @@ interface MapboxAudioGuidanceServices {
         language: String,
     ): MapboxVoiceInstructionsPlayer
 
-    fun mapboxVoiceInstructions(mapboxNavigation: MapboxNavigation): MapboxVoiceInstructions
+    fun mapboxVoiceInstructions(): MapboxVoiceInstructions
 }